+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Changelog
 0.3.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add `get_field_type_map` to `ServiceDirective`. Since Sphinx 4.0 it is required.
+  See https://github.com/sphinx-doc/sphinx/pull/7416
 
 
 0.3.1 (2018-04-21)